+++ ECG_Solution/src/Camera.hpp
@@ -19,7 +19,7 @@ private:
     bool backfaceCulling;
 
 public:
-    Camera(float fov, float height, float width, float zNear, float zFar);
+    Camera(float fov, int height, int width, float zNear, float zFar);
     glm::mat4 ViewProjMatrix();
     void translate(glm::vec3 trans);
     void rotate(glm::vec3 rot);
